+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.2.0 – 2024.11.?
 
+### Added
+
+- Developer experience: added example `/examples/local-app-workflows` for demonstrate work
+  with [workflows](https://apidocs.bitrix24.com/api-reference/bizproc/index.html).
+
 ### Fixed
 
 - Fixed errors in `Bitrix24\SDK\Services\Workflows\Common\WorkflowDocumentId`,